backend/app/schemas/auth.py
from pydantic import BaseModel, EmailStr

class RegisterRequest(BaseModel):
    email: EmailStr
    password: str
    tenant_name: str

class LoginRequest(BaseModel):
    email: EmailStr
    password: str