backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
cryptography
python-dotenv
passlib[bcrypt]
bcrypt==4.0.1
email-validator
python-jose[cryptography]
boto3